services/redaction/requirements.txt
presidio-analyzer==2.2.358
presidio-anonymizer==2.2.358
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-multipart==0.0.20
spacy==3.8.3
redis==5.2.1
requests==2.32.3
pytesseract==0.3.10
Pillow==9.5.0
pdf2image==1.16.3
